# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handlers.AddRoleRequest:
    properties:
      role:
        example: hr
        type: string
    required:
    - role
    type: object
  wowrack-recruitment_internal_dto.CreateDepartmentRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  wowrack-recruitment_internal_dto.UpdateDepartmentRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@albyeicorp.com
    name: Albyei Corp Engineering
  description: API untuk sistem rekrutmen Wowrack/Albyei Corp. Mendukung multi-role
    (Admin, HR, Hiring Manager, Candidate).
  termsOfService: http://swagger.io/terms/
  title: Wowrack Recruitment API
  version: "1.0"
paths:
...
  /admin/users/{id}/roles:
    get:
      description: Get all roles assigned to a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get user roles
      tags:
      - Admin Roles
    post:
      consumes:
      - application/json
      description: Assign a new role to a user (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.AddRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Add role to user
      tags:
      - Admin Roles
  /admin/users/{id}/roles/{role}:
    delete:
      description: Remove a specific role from a user (admin only). Cannot remove
        the last role.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role to remove
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Remove role from user
      tags:
      - Admin Roles
...
  /hr/departments:
    get:
      description: Get list of all departments (HR only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get all departments
      tags:
      - Departments
    post:
      consumes:
      - application/json
      description: Create a new department (HR only)
      parameters:
      - description: Department name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/wowrack-recruitment_internal_dto.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create department
      tags:
      - Departments
  /hr/departments/{id}:
    delete:
      description: Delete a department by ID (HR only)
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete department
      tags:
      - Departments
    get:
      description: Get a single department by its ID (HR only)
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get department by ID
      tags:
      - Departments
    put:
      consumes:
      - application/json
      description: Update an existing department name (HR only)
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Department data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/wowrack-recruitment_internal_dto.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update department
      tags:
      - Departments
...
  /jobs/{slug}/apply:
    post:
      consumes:
      - multipart/form-data
      description: Submit a job application with CV file upload (multipart/form-data)
      parameters:
      - description: Job slug
        in: path
        name: slug
        required: true
        type: string
      - description: Full name
        in: formData
        name: name
        required: true
        type: string
      - description: Email address
        in: formData
        name: email
        required: true
        type: string
      - description: Phone number
        in: formData
        name: phone
        required: true
        type: string
      - description: Terms acceptance
        in: formData
        name: terms_accepted
        required: true
        type: boolean
      - description: WhatsApp number (08xx format)
        in: formData
        name: whatsapp_number
        required: true
        type: string
      - description: City of domicile
        in: formData
        name: domicile_city
        required: true
        type: string
      - description: Province of domicile
        in: formData
        name: domicile_province
        required: true
        type: string
      - description: University name
        in: formData
        name: university
        type: string
      - description: CV file (PDF, max 5MB)
        in: formData
        name: cv
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Apply for a job
      tags:
      - Applications
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"